Cambios generales

+++ src/commons/recaudo/screens/GeneradorDocumentos.tsx
@@ -22,9 +22,7 @@ import SaveIcon from '@mui/icons-material/Save';
 import EditIcon from '@mui/icons-material/Edit';
 import FeedIcon from '@mui/icons-material/Feed';
 import CloseIcon from '@mui/icons-material/Close';
-import CloudUploadIcon from '@mui/icons-material/CloudUpload';
 import UpdateIcon from '@mui/icons-material/Update';
-import { AlertaDocumento } from './AlertaDocumento';
 import { FormularioGenerador } from '../components/generadorDocs/FormularioGenerador';
 import { VisorDocumentos } from '../components/GeneradorDocumentos/VisorDocumentos';
 import { resetBandejaDeTareas } from '../../gestorDocumental/bandejaDeTareas/toolkit/store/BandejaDeTareasStore';
@@ -54,15 +52,16 @@ export const GeneradorDocumentos: React.FC = () => {
   const [updateDocument, setUpdateDocument] = useState(false);
   const [sendTemplate, setSendTemplate] = useState(false);
   const [uplockFirma, setUplockFirma] = useState(false);
+  const [noneFirma, setNoneFirma] = useState(false);
   const [isNewData, setIsNewData] = useState(false);
   const [hasConsecutivo, setHasConsecutivo] = useState(false);
   const [hasRadicado, setHasRadicado] = useState(false);
   const [hasPersona, setHasPersona] = useState(false);
 
-  const [dataPQRS, setDataPQRS] = useState<any>({}); // Datos para llenar el documento
+  const [contentData, setContentData] = useState<any>({}); // Datos para llenar el documento
   const [matchingData, setMatchingData] = useState<any>({}); // Datos para llenar el documento
 
-  const [personaSelected, setpersona] = useState<string[]>([]);
+  const [personaSelected, setpersona] = useState<any[]>([]);
   const [perfilselet, setperfilselet] = useState<string[]>([]); // Asumiendo que es un string
   const urlBase = baseURL.replace("/api/", "");
 
@@ -74,6 +73,7 @@ export const GeneradorDocumentos: React.FC = () => {
     }else{
       setHasPersona(false);
     }
+    console.log(personaSelected);
   }, [personaSelected])
 
   useEffect(() => {
@@ -85,18 +85,29 @@ export const GeneradorDocumentos: React.FC = () => {
   const currentElement = useSelector((state: any) => state.BandejaTareasSlice.currentElementBandejaTareasPqrsdfYTramitesYOtrosYOpas);
 
   useEffect(() => {
-    if(currentElement && currentElement.documento){
-      removeFile()
-      if(currentElement.documento.id_consecutivo_tipologia){
-        setCurrentBorrador(currentElement.documento);
-        setHasConsecutivo(true);
+    if(currentElement){
+      if(currentElement.documento){
+        removeFile()
+        if(currentElement.documento.id_consecutivo_tipologia){
+          setCurrentBorrador(currentElement.documento);
+          setHasConsecutivo(true);
+          setUplockFirma(true);
+          if(!currentElement?.asignaciones?.firma){
+            setUplockFirma(false);
+          }
+        }
+        setFile(`${urlBase}${currentElement.documento.archivos_digitales.ruta_archivo}`)
       }
-      setFile(`${urlBase}${currentElement.documento.archivos_digitales.ruta_archivo}`)
-    }
 
-    if(currentElement && currentElement.id_pqrsdf){
-      removeFile()
-      search_pqrt(currentElement.id_pqrsdf);
+      if(currentElement.id_pqrsdf){
+        removeFile()
+        search_pqrt(currentElement.id_pqrsdf);
+      }
+
+      if(currentElement.id_tramite && !currentElement?.id_pqrsdf && !currentElement?.documento){
+        removeFile()
+        search_opa(currentElement.id_tramite);
+      }
     }
     console.log(currentElement);
   }, [currentElement])
@@ -108,12 +119,27 @@ export const GeneradorDocumentos: React.FC = () => {
       console.log(response.data);
       if(response?.data?.data){
         control_success('Datos cargados correctamente, elige la plantalla correspondiente');
-        setDataPQRS(response.data.data);
+        setContentData(response.data.data);
       }
     } catch (error: any) {
       control_error(error.response.data.detail);
     }
   }
+
+  const search_opa = async (id: string) => {
+    try {
+      const url = `gestor/pqr/get_opa-panel/${id}/`;
+      const response = await api.get(url);
+      if(response?.data?.data){
+        control_success('Datos cargados correctamente, elige la plantalla correspondiente');
+        setContentData(response.data.data);
+      }
+    } catch (error: any) {
+      control_error(error.response.data.detail);
+    }
+  }
+
+
 
   const [radicado_selected, set_radicado_selected] = useState('');
 
@@ -155,7 +181,11 @@ export const GeneradorDocumentos: React.FC = () => {
     target: { value: SetStateAction<string> };
   }) => {
     const id_plantilla = event.target.value;
+    setUpdateBorrador(false)
+    setSendTemplate(false)
     setHasConsecutivo(false);
+    set_radicado_selected('');
+    setUplockFirma(false);
     setIdPlantilla(id_plantilla)
     const currentPlantilla = plantillas.find(plantilla => plantilla.id_plantilla_doc === id_plantilla);
     if(currentPlantilla){
@@ -244,18 +274,27 @@ export const GeneradorDocumentos: React.FC = () => {
   useEffect(() => {
     if(plantillaSeleccionada?.archivos_digitales){
       removeFile()
-      setUpdateBorrador(false)
-      setSendTemplate(false)
       const url = baseURL.replace("/api/", "");
       setFile(`${url}${plantillaSeleccionada.archivos_digitales.ruta_archivo}`)
       let variablesFiltradas = plantillaSeleccionada.variables.filter((variable: string) => variable !== 'consecutivo' && variable !== 'radicado' && variable !== 'fecha_radicado');
       let newMatchingData: any = {};
-      if (dataPQRS?.id_PQRSDF) {
-        const keysdataPQRS = Object.keys(dataPQRS.info_persona_titular);
-        console.log(keysdataPQRS);
+      if (contentData?.id_PQRSDF) {
+        const keyscontentData = Object.keys(contentData.info_persona_titular);
         variablesFiltradas = variablesFiltradas.filter((variable: string) => {
-          if (keysdataPQRS.includes(variable) && dataPQRS.info_persona_titular[variable]) {
-            newMatchingData[variable] = dataPQRS.info_persona_titular[variable];
+          if (keyscontentData.includes(variable) && contentData.info_persona_titular[variable]) {
+            newMatchingData[variable] = contentData.info_persona_titular[variable];
+            return false;
+          }
+          return true;
+        });
+      }
+
+      if(contentData?.id_tramite && !contentData?.id_PQRSDF){
+        const keyscontentData = Object.keys(contentData.info_persona_titular);
+        console.log(keyscontentData);
+        variablesFiltradas = variablesFiltradas.filter((variable: string) => {
+          if (keyscontentData.includes(variable) && contentData.info_persona_titular[variable]) {
+            newMatchingData[variable] = contentData.info_persona_titular[variable];
             return false;
           }
           return true;
@@ -264,7 +303,7 @@ export const GeneradorDocumentos: React.FC = () => {
       setMatchingData(newMatchingData);
       setVariablesPlantilla(variablesFiltradas);
     }
-  }, [plantillaSeleccionada, dataPQRS]);
+  }, [plantillaSeleccionada, contentData]);
 
   useEffect(() => {
     if (plantillaSeleccionada?.archivos_digitales && hasValue(matchingData)) {
@@ -292,7 +331,6 @@ export const GeneradorDocumentos: React.FC = () => {
       plantilla: idPlantilla,
       payload: {...data, ...matchingData},
     };
-    console.log(sendData, variablesPlantilla);
     if(hasValue(data) || hasValue(matchingData)){
 
       if(updateBorrador){
@@ -312,7 +350,10 @@ export const GeneradorDocumentos: React.FC = () => {
       }
 
       if(updateDocument){
-        if(!hasRadicado) sendData.cod_tipo_radicado = radicado_selected;
+        if(!hasRadicado) {
+          sendData.cod_tipo_radicado = radicado_selected;
+          setHasRadicado(true);
+        }
         sendData.variable = 'A'
         sendData.id_consecutivo = currentBorrador?.id_consecutivo_tipologia;
         setHasConsecutivo(true);
@@ -345,10 +386,24 @@ export const GeneradorDocumentos: React.FC = () => {
   const sendDocument = async () => {
     if(personaSelected.length){
       let allSuccess = true;
-      for (const id_persona_asignada of personaSelected) {
+      if(id_persona){
+        const exists = personaSelected.some(persona => persona.id === id_persona);
+        if (!exists) {
+          const body = {
+            id_consecutivo: currentBorrador?.id_consecutivo_tipologia,
+            id_persona_asignada: id_persona,
+            firma: true
+          }
+          const success = await createAsignacionDocumento(body);
+          if (!success) allSuccess = false;
+        }
+      }
+
+      for (const persona of personaSelected) {
         const body = {
-          id_consecutivo: currentBorrador?.id_consecutivo_tipologia,
-          id_persona_asignada,
+            id_consecutivo: currentBorrador?.id_consecutivo_tipologia,
+            id_persona_asignada: persona.id,
+            firma: persona.require_firma
         }
         const success = await createAsignacionDocumento(body);
         if (!success) allSuccess = false;
@@ -357,7 +412,7 @@ export const GeneradorDocumentos: React.FC = () => {
         setUplockFirma(true);
         if (personaSelected.length == 1) control_success('El documento se envió correctamente');
         if (personaSelected.length > 1) control_success('Todos los documentos se enviaron correctamente');
-        cleanTemplate();
+        // cleanTemplate();
       } else {
         if (personaSelected.length == 1) control_error('Ocurrio un error al enviar el documento');
         if (personaSelected.length > 1) control_error('Hubo un error al enviar algunos documentos');
@@ -366,7 +421,8 @@ export const GeneradorDocumentos: React.FC = () => {
   }
 
   const cleanTemplate = () => {
-    setPlantillaSeleccionada(null);
+    setPlantillaSeleccionada('');
+    setVariablesPlantilla([]);
     setFile('');
     setSendTemplate(false);
     setUpdateBorrador(false);
@@ -562,9 +618,12 @@ export const GeneradorDocumentos: React.FC = () => {
         }}
       >
         <VisorDocumentos
+          clean_template={cleanTemplate}
           file={file}
           current_borrador={currentBorrador}
-          uplock_firma={hasConsecutivo}
+          none_firma={noneFirma}
+          uplock_firma={uplockFirma}
+          set_uplock_firma={setUplockFirma}
         />
 
       </Grid>}
